move notebook_utils.py to utils; update imports
