+++ openmdao/utils/variable_table.py
@@ -9,11 +9,7 @@ from io import TextIOBase
 import numpy as np
 
 from openmdao.core.constants import _DEFAULT_OUT_STREAM
-from openmdao.core.notebook_utils import notebook, tabulate
-try:
-    from IPython.display import display, HTML
-except ImportError:
-    display = HTML = None
+from openmdao.utils.notebook_utils import notebook, tabulate, display, HTML
 
 column_widths = {
     'value': 20,
